[refactor] 안쓰는 web로그인 클래스 제거2
