+++ src/main/java/com/kustaurant/kustaurant/common/user/controller/api/MypageApiController.java
@@ -1,7 +1,7 @@
 package com.kustaurant.kustaurant.common.user.controller.api;
 
 import com.kustaurant.kustaurant.common.notice.domain.NoticeDTO;
-import com.kustaurant.kustaurant.common.user.domain.*;
+import com.kustaurant.kustaurant.common.user.controller.api.response.*;
 import com.kustaurant.kustaurant.global.apiUser.customAnno.JwtToken;
 import com.kustaurant.kustaurant.common.user.service.MypageApiService;
 import io.swagger.v3.oas.annotations.Operation;
