+++ src/main/java/com/kustaurant/kustaurant/common/user/controller/api/response/MypageMainDTO.java
@@ -1,4 +1,4 @@
-package com.kustaurant.kustaurant.common.user.domain;
+package com.kustaurant.kustaurant.common.user.controller.api.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
