+++ src/main/java/com/kustaurant/kustaurant/common/user/controller/api/response/MypagePostCommentDTO.java
@@ -1,4 +1,4 @@
-package com.kustaurant.kustaurant.common.user.domain;
+package com.kustaurant.kustaurant.common.user.controller.api.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,13 +7,11 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class MypagePostDTO {
+public class MypagePostCommentDTO {
     private Integer postId;
     private String postCategory;
     private String postTitle;
-    private String postImgUrl;
-    private String postBody;
-    private Integer likeCount;
-    private Integer commentCount;
+    private String postcommentBody;
+    private Integer commentlikeCount;
     private String timeAgo;
 }
