+++ src/main/java/com/kustaurant/kustaurant/global/webUser/UserRole.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 public enum UserRole {
     ADMIN("ROLE_ADMIN"),
     USER("ROLE_USER");
+
     UserRole(String value){
         this.value = value;
     }
